+++ sentry-spring-boot-starter-jakarta/src/main/java/io/sentry/spring/boot/jakarta/SentryAutoConfiguration.java
@@ -119,9 +119,6 @@ public class SentryAutoConfiguration {
           BuildConfig.SENTRY_SPRING_BOOT_JAKARTA_SDK_NAME + "/" + BuildConfig.VERSION_NAME);
       options.setSdkVersion(createSdkVersion(options));
       addPackageAndIntegrationInfo();
-      if (options.getTracesSampleRate() == null && options.getEnableTracing() == null) {
-        options.setTracesSampleRate(0.0);
-      }
       // Spring Boot sets ignored exceptions in runtime using reflection - where the generic
       // information is lost
       // its technically possible to set non-throwable class to `ignoredExceptionsForType` set
@@ -240,7 +237,6 @@ public class SentryAutoConfiguration {
       }
 
       @Bean
-      @Conditional(SentryTracingCondition.class)
       @ConditionalOnMissingBean(name = "sentryTracingFilter")
       public FilterRegistrationBean<SentryTracingFilter> sentryTracingFilter(
           final @NotNull IHub hub, final @NotNull TransactionNameProvider transactionNameProvider) {