+++ sentry/src/main/java/io/sentry/TransactionContext.java
@@ -3,13 +3,18 @@ package io.sentry;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.TransactionNameSource;
 import io.sentry.util.Objects;
+import java.util.List;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public final class TransactionContext extends SpanContext {
-  private final @NotNull String name;
-  private final @NotNull TransactionNameSource transactionNameSource;
+  private static final @NotNull String DEFAULT_NAME = "<unlabeled transaction>";
+  private static final @NotNull TransactionNameSource DEFAULT_NAME_SOURCE =
+      TransactionNameSource.CUSTOM;
+  private static final @NotNull String DEFAULT_OPERATION = "default";
+  private @NotNull String name;
+  private @NotNull TransactionNameSource transactionNameSource;
   private @Nullable TracesSamplingDecision parentSamplingDecision;
   private @Nullable Baggage baggage;
   private @NotNull Instrumenter instrumenter = Instrumenter.SENTRY;
@@ -20,69 +25,48 @@ public final class TransactionContext extends SpanContext {
    * @param name - the transaction name
    * @param operation - the operation
    * @param sentryTrace - the sentry-trace header
+   * @deprecated use {@link Sentry#continueTrace(String, List)} and setters for name and operation
+   *     here instead.
    * @return the transaction contexts
    */
+  @Deprecated
   public static @NotNull TransactionContext fromSentryTrace(
       final @NotNull String name,
       final @NotNull String operation,
       final @NotNull SentryTraceHeader sentryTrace) {
-    return fromSentryTrace(name, TransactionNameSource.CUSTOM, operation, sentryTrace, null, null);
-  }
-
-  /**
-   * Creates {@link TransactionContext} from sentry-trace header.
-   *
-   * @param name - the transaction name
-   * @param transactionNameSource - source of the transaction name
-   * @param operation - the operation
-   * @param sentryTrace - the sentry-trace header
-   * @return the transaction contexts
-   */
-  @ApiStatus.Internal
-  public static @NotNull TransactionContext fromSentryTrace(
-      final @NotNull String name,
-      final @NotNull TransactionNameSource transactionNameSource,
-      final @NotNull String operation,
-      final @NotNull SentryTraceHeader sentryTrace) {
-    @Nullable Boolean parentSampled = sentryTrace.isSampled();
-    return new TransactionContext(
-        name,
-        operation,
-        sentryTrace.getTraceId(),
-        new SpanId(),
-        transactionNameSource,
-        sentryTrace.getSpanId(),
-        parentSampled == null ? null : new TracesSamplingDecision(parentSampled),
-        null);
-  }
-
-  /**
-   * Creates {@link TransactionContext} from sentry-trace header.
-   *
-   * @param name - the transaction name
-   * @param transactionNameSource - source of the transaction name
-   * @param operation - the operation
-   * @param sentryTrace - the sentry-trace header
-   * @param baggage - the baggage header
-   * @return the transaction contexts
-   */
-  @ApiStatus.Internal
-  public static @NotNull TransactionContext fromSentryTrace(
-      final @NotNull String name,
-      final @NotNull TransactionNameSource transactionNameSource,
-      final @NotNull String operation,
-      final @NotNull SentryTraceHeader sentryTrace,
-      final @Nullable Baggage baggage,
-      final @Nullable SpanId spanId) {
     @Nullable Boolean parentSampled = sentryTrace.isSampled();
     TracesSamplingDecision samplingDecision =
         parentSampled == null ? null : new TracesSamplingDecision(parentSampled);
+
+    TransactionContext transactionContext =
+        new TransactionContext(
+            sentryTrace.getTraceId(),
+            new SpanId(),
+            sentryTrace.getSpanId(),
+            samplingDecision,
+            null);
+
+    transactionContext.setName(name);
+    transactionContext.setTransactionNameSource(TransactionNameSource.CUSTOM);
+    transactionContext.setOperation(operation);
+
+    return transactionContext;
+  }
+
+  @ApiStatus.Internal
+  public static TransactionContext fromPropagationContext(
+      final @NotNull PropagationContext propagationContext) {
+    @Nullable Boolean parentSampled = propagationContext.isSampled();
+    TracesSamplingDecision samplingDecision =
+        parentSampled == null ? null : new TracesSamplingDecision(parentSampled);
+
+    @Nullable Baggage baggage = propagationContext.getBaggage();
 
     if (baggage != null) {
       baggage.freeze();
 
       Double sampleRate = baggage.getSampleRateDouble();
-      Boolean sampled = parentSampled != null ? parentSampled.booleanValue() : true;
+      Boolean sampled = parentSampled != null ? parentSampled.booleanValue() : false;
       if (sampleRate != null) {
         samplingDecision = new TracesSamplingDecision(sampled, sampleRate);
       } else {
@@ -90,15 +74,10 @@ public final class TransactionContext extends SpanContext {
       }
     }
 
-    final @NotNull SpanId spanIdToUse = spanId == null ? new SpanId() : spanId;
-
     return new TransactionContext(
-        name,
-        operation,
-        sentryTrace.getTraceId(),
-        spanIdToUse,
-        transactionNameSource,
-        sentryTrace.getSpanId(),
+        propagationContext.getTraceId(),
+        propagationContext.getSpanId(),
+        propagationContext.getParentSpanId(),
         samplingDecision,
         baggage);
   }
@@ -150,18 +129,15 @@ public final class TransactionContext extends SpanContext {
 
   @ApiStatus.Internal
   public TransactionContext(
-      final @NotNull String name,
-      final @NotNull String operation,
       final @NotNull SentryId traceId,
       final @NotNull SpanId spanId,
-      final @NotNull TransactionNameSource transactionNameSource,
       final @Nullable SpanId parentSpanId,
       final @Nullable TracesSamplingDecision parentSamplingDecision,
       final @Nullable Baggage baggage) {
-    super(traceId, spanId, operation, parentSpanId, null);
-    this.name = Objects.requireNonNull(name, "name is required");
+    super(traceId, spanId, DEFAULT_OPERATION, parentSpanId, null);
+    this.name = DEFAULT_NAME;
     this.parentSamplingDecision = parentSamplingDecision;
-    this.transactionNameSource = transactionNameSource;
+    this.transactionNameSource = DEFAULT_NAME_SOURCE;
     this.baggage = baggage;
   }
 
@@ -215,5 +191,13 @@ public final class TransactionContext extends SpanContext {
 
   public void setInstrumenter(final @NotNull Instrumenter instrumenter) {
     this.instrumenter = instrumenter;
+  }
+
+  public void setName(final @NotNull String name) {
+    this.name = Objects.requireNonNull(name, "name is required");
+  }
+
+  public void setTransactionNameSource(final @NotNull TransactionNameSource transactionNameSource) {
+    this.transactionNameSource = transactionNameSource;
   }
 }